requirements.txt
boto3
streamlit
pandas
python-dotenv
openpyxl
fpdf
altair
requests
plotly